adding create tagihan, generate tagihan, bayar tagihan, confirm payment, tagihan history

+++ lib/widgets/checklist_tagihan_item.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
 import 'package:si_warga/pages/edit_tagihan.dart';
 
 class ChecklistTagihanItem extends StatefulWidget {
@@ -7,12 +8,14 @@ class ChecklistTagihanItem extends StatefulWidget {
   final int value;
   final String tagihanId;
   final Map<String, dynamic> tagihanData;
+  // final String tenggat;
   const ChecklistTagihanItem({
     super.key,
     required this.title,
     required this.value,
     required this.tagihanId,
     required this.tagihanData,
+    // required this.tenggat,
   });
 
   @override
@@ -41,6 +44,37 @@ Future<void> hapusTagihan(String tagihanId) async {
 }
 
 class _ChecklistTagihanItemState extends State<ChecklistTagihanItem> {
+  final formatter = NumberFormat.currency(
+    locale: 'id_ID',
+    symbol: 'Rp ',
+    decimalDigits: 0,
+  );
+
+  // late final String formattedTenggat;
+
+  // @override
+  // void initState() {
+  //   super.initState();
+  //   formattedTenggat = _formatTenggat(widget.tenggat);
+  // }
+
+  // String _formatTenggat(dynamic tenggat) {
+  //   try {
+  //     if (tenggat is Timestamp) {
+  //       return DateFormat('dd MMM yyyy').format(tenggat.toDate());
+  //     } else if (tenggat is String) {
+  //       final parsedDate = DateTime.parse(tenggat);
+  //       return DateFormat('dd MMM yyyy').format(parsedDate);
+  //     } else if (tenggat is DateTime) {
+  //       return DateFormat('dd MMM yyyy').format(tenggat);
+  //     } else {
+  //       return '-';
+  //     }
+  //   } catch (e) {
+  //     return '-';
+  //   }
+  // }
+
   @override
   Widget build(BuildContext context) {
     return Padding(
@@ -59,7 +93,8 @@ class _ChecklistTagihanItemState extends State<ChecklistTagihanItem> {
           Row(
             children: [
               Text(
-                widget.value.toString(),
+                // widget.value.toString(),
+                formatter.format(widget.value),
                 style: TextStyle(fontWeight: FontWeight.w500, fontSize: 16),
               ),
               SizedBox(width: 10),
@@ -81,7 +116,41 @@ class _ChecklistTagihanItemState extends State<ChecklistTagihanItem> {
               SizedBox(width: 10),
               GestureDetector(
                 onTap: () async {
-                  await hapusTagihan(widget.tagihanId);
+                  // await hapusTagihan(widget.tagihanId);
+                  final confirmDelete = await showDialog<bool>(
+                    context: context,
+                    builder:
+                        (context) => AlertDialog(
+                          title: Text('Konfirmasi Hapus Tagihan'),
+                          content: Text(
+                            'Apakah Anda yakin ingin menghapus tagihan?',
+                          ),
+                          actions: [
+                            TextButton(
+                              onPressed: () => Navigator.of(context).pop(false),
+                              child: Text(
+                                'Batal',
+                                style: TextStyle(color: Color(0xFF777777)),
+                              ),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                // hapusTagihan(widget.tagihanId);
+                                Navigator.of(context).pop(true);
+                              },
+                              child: Text(
+                                'Hapus',
+                                style: TextStyle(
+                                  color: Color.fromARGB(255, 255, 42, 42),
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                  );
+                  if (confirmDelete == true) {
+                    await hapusTagihan(widget.tagihanId);
+                  }
                 },
                 child: Icon(
                   Icons.delete_outline_rounded,
